requirements_docs.txt
recommonmark==0.7.1
matplotlib==3.5.2
imageio==2.21.1
imageio-ffmpeg==0.4.7
pyvista==0.36.1
numpydoc==1.4.0
Sphinx==5.1.1
sphinx-autobuild==2021.3.14
sphinxcontrib-websupport==1.2.4
pytest-sphinx==0.4.0
sphinx-notfound-page==0.8
sphinx-copybutton==0.5.0
sphinx-gallery==0.11.0
sphinx-autodoc-typehints==1.19.2
ansys-sphinx-theme==0.4.2
pypandoc==1.8.1
nbsphinx==0.8.9
ipywidgets==7.7.1
joblib==1.1.0
scikit-learn==1.1.2
